backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI/ML models - Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.51.3
# Note: diffusers should be installed from GitHub
# pip install git+https://github.com/huggingface/diffusers

# Image processing
Pillow==10.1.0
python-pptx==0.6.23
numpy==1.26.2

# Async and utilities
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis (optional)
redis==5.0.1
hiredis==2.2.3

# Monitoring and logging
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
